Move debugging protocol types to Workspaces
