+++ src/Razor/src/Microsoft.CodeAnalysis.Razor.Workspaces/Protocol/Debugging/RazorBreakpointSpanParams.cs
@@ -1,11 +1,14 @@
 ﻿// Copyright (c) .NET Foundation. All rights reserved.
 // Licensed under the MIT license. See License.txt in the project root for license information.
 
+using System;
 using Microsoft.VisualStudio.LanguageServer.Protocol;
 
-namespace Microsoft.AspNetCore.Razor.LanguageServer.Protocol;
+namespace Microsoft.CodeAnalysis.Razor.Protocol.Debugging;
 
-internal class RazorBreakpointSpanResponse
+internal class RazorBreakpointSpanParams
 {
-    public required Range Range { get; init; }
+    public required Uri Uri { get; init; }
+
+    public required Position Position { get; init; }
 }
